\name{plotMakefile-package}
\alias{plotMakefile-package}
\alias{plotMakefile}
\docType{package}
\title{Parse and plot Makefiles}
\description{A Makefile encodes a dependency tree between files in a project,
which can be visualized using plotMakefile("path/to/Makefile").}
\details{
\tabular{ll}{Package: \tab plotMakefile\cr
Maintainer: \tab Toby Dylan Hocking <tdhock5@gmail.com>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2014.2.13\cr
License: \tab GPL-3\cr
Title: \tab Parse and plot Makefiles\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


