\name{sortValues}
\alias{sortValues}
\title{sortValues}
\description{manually sort to avoid edge crossings.}
\usage{sortValues(...)}
\arguments{
  \item{\dots}{file names in the order that they should appear on the plot.}
}

\value{a function that can be used as the sort.fun argument to
plotmakefile.}

\author{Toby Dylan Hocking}





