\name{str_match_all_perl}
\alias{str_match_all_perl}
\title{str match all perl}
\description{Parse several occurances of pattern from each of several strings
using (named) capturing regular expressions, returning a list of
matrices (with column names).}
\usage{str_match_all_perl(string, pattern)}
\arguments{
  \item{string}{
}
  \item{pattern}{
}
}



\author{Toby Dylan Hocking}





